mujoco/requirements.txt
absl-py==0.6.1
astor==0.7.1
atari-py==0.1.1
better-exceptions==0.1.8
bleach==1.5.0
Box2D==2.3.2
Box2D-kengz==2.3.3
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.2
cycler==0.10.0
Cython==0.27.3
dask==1.0.0
decorator==4.1.2
dm-control==0.0.0
entrypoints==0.2.3
enum34==1.1.6
future==0.16.0
futures==3.1.1
gast==0.2.0
glfw==1.7.0
grpcio==1.17.0
gym==0.10.8
h5py==2.8.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
imgcat==0.2.0
ipdb==0.11
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lockfile==0.12.2
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.1
moviepy==0.2.3.5
mpi4py==3.0.0
mujoco-py==1.50.1.65
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.2.1
numpy==1.14.1
olefile==0.44
pachi-py==0.0.21
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pbr==3.1.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.5.2
pycparser==2.19
pygco==0.0.1
pyglet==1.3.0
Pygments==2.2.0
PyOpenGL==3.1.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==1.0.1
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.14.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
stevedore==1.28.0
tensorboard==1.12.1
tensorflow-gpu==1.12.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.7
testpath==0.3.1
tf-video-summary==0.1.0
Theano==1.0.1
toolz==0.9.0
tornado==4.5.2
tqdm==4.29.1
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
Wand==0.4.4
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.0.8
